requirements.txt
cv_bridge==1.16.2
matplotlib==3.10.7
message_filters==1.17.0
moveit_commander==1.1.16
numpy==2.3.4
open3d==0.19.0
opencv_python==4.12.0.88
Pillow==12.0.0
pyrealsense2==2.56.5.9235
PyYAML==6.0.2
rospy==1.17.0
scikit_learn==1.7.2
scipy==1.16.2
sensor_msgs==1.13.1
setuptools==58.1.0
supervision==0.26.1
tf==1.13.3
torch==2.5.1
empy==3.3.4 # add this if you cannot build ros pkg with catkin_make
